code review round 10: fix type mismatch in wal_relay logging
Round 10 fixes:

Type safety:
- wal_relay.cc: Fix type mismatch in SPDLOG_INFO format - lines.size()
  (size_t) - relayed (int64_t) could cause sign-compare warning.
  Cast lines.size() to int64_t explicitly.

+++ src/wal/wal_relay.cc
@@ -385,7 +385,7 @@ int64_t WalRelay::processWalFile(const std::string& file_path) {
         SPDLOG_INFO("WalRelay: {} of {} alerts relayed from '{}', "
                     "remaining {} kept in file",
                     relayed, lines.size(), file_path,
-                    lines.size() - relayed);
+                    static_cast<int64_t>(lines.size()) - relayed);
     }
 
     return relayed;
